requirements.txt
faiss-cpu
google-cloud-aiplatform
pandas
tqdm
plantuml
